compare product

+++ application/controllers/Compare_controller.php
@@ -13,16 +13,18 @@ class Compare_controller extends Home_Core_Controller
 	{
 		$arr_product_id = $this->session->userdata("compared_products_id");
 		$product_quantity = $this->session->userdata("compared_products_quantity");
-		$arr_id = [17, 18];
+		// $arr_id = [17];
 		$product_quantity = 2;
+		// $this->session->set_userdata("compared_products_id", array(19));
 
 		$data['title'] = "Perbandingan Produk";
-		$data['description'] = "membandingkan produk" . " - " . $this->app_name;
+		$data['description'] = "Membandingkan produk" . " - " . $this->app_name;
 		$data['keywords'] = "compare-product" . "," . $this->app_name;
 		$data['arr_payment_source'] = array("Dana Bos");
 		$data['products'] = $this->compare_model->get_compared_products_by_id($arr_product_id, $product_quantity);
 		$data['product_quantity'] = $product_quantity;
 		$data['list_vendors'] = $this->compare_model->get_all_vendors();
+		$data['temp_vendor'] = $this->session->userdata('temp_vendor');
 
 		$this->load->view('partials/_header', $data);
 		$this->load->view('compare/compare', $data);
@@ -41,8 +43,9 @@ class Compare_controller extends Home_Core_Controller
 	public function add_compared_product()
 	{
 		$compared_products = $this->session->userdata("compared_products_id");
-		$compared_products[] = $this->input->get("product_id");
+		array_push($compared_products, $this->input->get("product_id"));
 		$this->session->set_userdata("compared_products_id", $compared_products);
+		$this->session->unset_userdata("temp_vendor");
 		redirect(base_url('compare'));
 	}
 
@@ -50,8 +53,18 @@ class Compare_controller extends Home_Core_Controller
 	{
 		$compared_products = $this->session->userdata("compared_products_id");
 		$deleted_compared_product = $this->input->get("compared_product_id");
-		$compared_products = $this->compare_model->delete_product($compared_products, $deleted_compared_product);
-		$this->session->set_userdata("compared_products_id", $compared_products);
+		if ($deleted_compared_product != null) {
+			$compared_products = $this->compare_model->delete_product($compared_products, $deleted_compared_product);
+			$this->session->set_userdata("compared_products_id", $compared_products);
+		}
+		$this->session->unset_userdata("temp_vendor");
 		redirect(base_url('compare'));
+	}
+
+	public function add_compared_vendor()
+	{
+		$vendor_id = $this->input->get("vendor_id");
+		$this->session->set_userdata("temp_vendor", get_user($vendor_id));
+		redirect('compare');
 	}
 }
